+++ OneUpMushroom.cpp
@@ -17,8 +17,8 @@ void OneUpMushroom::Paint()
 {
 	double left = m_ActPtr->GetPosition().x;
 	double top = m_ActPtr->GetPosition().y;
-	double srcCol = 2;
-	double srcRow = 12;
+	int srcCol = 2;
+	int srcRow = 12;
 
 	SpriteSheetManager::generalTilesPtr->Paint(left, top, srcCol, srcRow);
 }
